==> 0_organization_birth.R <==
###birth summary
setwd("C:/Users/MARCOS.ANTUNES/Documents")
load("df_0105.RData")
load("df_0610.RData")
load("df_1116.RData")
load("df_1723.RData")
gc()
df_0105$CONTADOR = NULL
df_0610$CONTADOR = NULL
df_1116$CONTADOR = NULL
df_1723$CONTADOR = NULL

df_0105$CODANOMAL = NULL
df_0610$CODANOMAL = NULL
df_1116$CODANOMAL = NULL
df_1723$CODANOMAL = NULL

df_0105$CODESTAB = NULL
df_0610$CODESTAB = NULL
df_1116$CODESTAB = NULL
df_1723$CODESTAB = NULL

df_0105$CODMUNNASC = NULL
df_0610$CODMUNNASC = NULL
df_1116$CODMUNNASC = NULL
df_1723$CODMUNNASC = NULL

df_0105$QTDFILMORT = NULL
df_0610$QTDFILMORT = NULL
df_1116$QTDFILMORT = NULL
df_1723$QTDFILMORT = NULL

df_0105$QTDFILVIVO = NULL
df_0610$QTDFILVIVO = NULL
df_1116$QTDFILVIVO = NULL
df_1723$QTDFILVIVO = NULL


df = rbind(df_0105,df_0610,df_1116,df_1723)
rm(df_0105,df_0610,df_1116,df_1723)

df$LOCNASC = NULL

save(df,file="df.RData",compression_level = 9)
rm(df)
